pyproject.toml
[project]
name = "itg-buddy-py"
version = "0.1.0"
description = "A discord bot for Dance Games at Berkeley"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.4.0",
    "itg-cli>=1.0.0",
    "python-dotenv>=1.0.1",
]
